Sat, Apr 22, 2023, 2:42 PM -04:00

+++ readability/readability.c
@@ -31,7 +31,7 @@ int main(void)
 
     //printf("Index value is %i\n", grade_level);
 
-    int calculate =  (65 / 14) * (100);
+    int calculate =  ((float) 65 / 14) * (100);
     printf("%i\n", calculate);
 
 }
